This will add and rename feature-item api to fetch
